Refs #13 - Moved all Django schemas in one module: with inheritance, code is smaller and fits in one file.
