+++ django_confit/schemas/__init__.py
@@ -1,4 +1,5 @@
 """Schemas for configuration validation."""
+from __future__ import absolute_import
 import warnings
 
 import django
@@ -9,7 +10,7 @@ from django_confit.utils.importlib import import_member
 def get_django_schema_class():
     """Return colander schema class for current (installed) Django version."""
     schema_import_path = \
-        'django_confit.schemas.django_{major}_{minor}' \
+        'django_confit.schemas.django' \
         '.Django{major}_{minor}_{micro}ConfigurationSchema' \
         .format(
             major=django.VERSION[0],
